Remove task-specific classes
