+++ mobile-browser-based-version/server/tasks.json
@@ -53,7 +53,9 @@
         "epochs": 10
       },
       "receivedMessagesThreshold": 1,
-      "dataType": "csv"
+      "dataType": "csv",
+      "inputColumns": ["PassengerId", "Age", "SibSp", "Parch", "Fare", "Pclass"],
+      "outputColumn": "Survived"
     }
   },
   {
@@ -86,7 +88,8 @@
       "dataType": "image",
       "IMAGE_H": 28,
       "IMAGE_W": 28,
-      "LABEL_LIST": ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9"]
+      "LABEL_LIST": ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9"],
+      "aggregateImagesById": false
     }
   },
   {
@@ -122,7 +125,8 @@
       "IMAGE_W": 224,
       "LABEL_LIST": ["COVID-Positive", "COVID-Negative"],
       "NUM_CLASSES": 2,
-      "dataType": "image"
+      "dataType": "image",
+      "aggregateImagesById": true
     }
   }
 ]
