+++ mobile-browser-based-version/src/task_definition/csv_task.js
@@ -1,16 +1,12 @@
 import * as d3 from 'd3';
 import * as tf from '@tensorflow/tfjs';
 import { checkData } from '../helpers/data_validation_script/helpers-csv-tasks';
+import { Task } from './task.js';
 
 /**
  * Dummy class to hold the Titanic Task information
  */
-export class TitanicTask {
-  constructor(taskId, displayInformation, trainingInformation) {
-    this.taskId = taskId;
-    this.displayInformation = displayInformation;
-    this.trainingInformation = trainingInformation;
-  }
+export class CsvTask extends Task{
   /**
    * This functions takes as input a file (of type File) uploaded by the reader and checks
    * if the said file meets the constraints requirements and if so prepare the training data.
@@ -32,31 +28,30 @@ export class TitanicTask {
     if (startTraining) {
       console.log('User File Validated. Start parsing.');
       console.log(headerCopied);
+
+      let originalHeaders = headers.map(element => element['userHeader'])
+      let inputColumns = this.trainingInformation.inputColumns;
+      let indices = Array.from({length: headers.length}, (x, i) => i)
+      let inputIndices = indices.filter(i => inputColumns.includes(originalHeaders[i]))
       let Xcsv = d3.csvParse(
         file.target.result,
         function(d) {
-          return [
-            +d[headerCopied[0]], // PassengerId
-            +d[headerCopied[1]], // Survived
-            d[headerCopied[3]] == 'male' ? 1 : 0, // Sex
-            +d[headerCopied[4]], // Age
-            +d[headerCopied[5]], // SibSp
-            +d[headerCopied[6]], // Parch
-            +d[headerCopied[8]], // Fare
-            +d[headerCopied[11]], // Pclass
-          ];
+          let result = inputIndices.map(i => +d[headerCopied[i]])
+          return result
         },
         function(error, rows) {
+          console.log(error)
           console.log(rows);
         }
       );
-
+      let yIdx = originalHeaders.indexOf(this.trainingInformation.outputColumn)
       let ycsv = d3.csvParse(
         file.target.result,
         function(d) {
-          return +d[headerCopied[1]];
+          return +d[headerCopied[yIdx]];
         },
         function(error, rows) {
+          console.log(error)
           console.log(rows);
         }
       );
@@ -68,23 +63,5 @@ export class TitanicTask {
       console.log('Cannot start training.');
     }
     return { accepted: startTraining, Xtrain: Xtrain, ytrain: ytrain };
-  }
-
-  /**
-   * Defintion of the model associated to the task
-   */
-  async createModel() {
-    // To be put in a titanic task specific model
-    let newModel = await tf.loadLayersModel(
-      'https://deai-313515.ew.r.appspot.com/tasks/' +
-        this.taskId +
-        '/model.json'
-    );
-    const savePathDb = 'indexeddb://working_'.concat(
-      this.trainingInformation.modelId
-    );
-
-    // only keep this here
-    await newModel.save(savePathDb);
   }
 }
